requirements.txt
d20>=1.1.0
discord-py-slash-command>=1.1.0
discord.py>=1.6.0
